requirements.txt
Django==2.2.2
django-mysql==3.2.0
django-filter-2.2.0
djangorestframework==3.10.2
django-oauth-toolkit-1.2.0
model-mommy==1.6.0
mysql==0.0.2
mysqlclient==1.4.2.post1
requests==2.22.0
Shapely==1.6.4.post2
urllib3==1.25.3
